requirements.txt
asynch==0.2.3
blinker==1.7.0
certifi==2024.2.2
cfgv==3.4.0
charset-normalizer==3.3.2
ciso8601==2.3.1
click==8.1.7
clickhouse-driver==0.2.7
clickhouse-sqlalchemy==0.3.0
distlib==0.3.8
filelock==3.13.1
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
flask-unittest==0.1.3
greenlet==3.0.3
identify==2.5.35
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
leb128==1.0.5
lz4==4.3.3
MarkupSafe==2.1.5
nodeenv==1.8.0
platformdirs==4.2.0
pre-commit==3.6.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
SQLAlchemy==2.0.27
typing_extensions==4.10.0
tzlocal==5.2
urllib3==2.2.1
virtualenv==20.25.1
Werkzeug==3.0.1
zstd==1.5.5.1